+++ lucene/core/src/java/org/apache/lucene/search/DocValuesFieldExistsQuery.java
@@ -17,108 +17,21 @@
 package org.apache.lucene.search;
 
 import java.io.IOException;
-import java.util.Objects;
-import org.apache.lucene.index.DocValues;
-import org.apache.lucene.index.DocValuesType;
 import org.apache.lucene.index.FieldInfo;
-import org.apache.lucene.index.IndexOptions;
-import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.index.LeafReader;
-import org.apache.lucene.index.LeafReaderContext;
-import org.apache.lucene.index.PointValues;
-import org.apache.lucene.index.Terms;
 
 /**
  * A {@link Query} that matches documents that have a value for a given field as reported by doc
  * values iterators.
+ *
+ * @deprecated Use {@link org.apache.lucene.search.FieldExistsQuery} instead.
  */
-public final class DocValuesFieldExistsQuery extends Query {
-
-  private final String field;
+@Deprecated
+public final class DocValuesFieldExistsQuery extends FieldExistsQuery {
 
   /** Create a query that will match documents which have a value for the given {@code field}. */
   public DocValuesFieldExistsQuery(String field) {
-    this.field = Objects.requireNonNull(field);
-  }
-
-  public String getField() {
-    return field;
-  }
-
-  @Override
-  public boolean equals(Object other) {
-    return sameClassAs(other) && field.equals(((DocValuesFieldExistsQuery) other).field);
-  }
-
-  @Override
-  public int hashCode() {
-    return 31 * classHash() + field.hashCode();
-  }
-
-  @Override
-  public String toString(String field) {
-    return "DocValuesFieldExistsQuery [field=" + this.field + "]";
-  }
-
-  @Override
-  public void visit(QueryVisitor visitor) {
-    if (visitor.acceptField(field)) {
-      visitor.visitLeaf(this);
-    }
-  }
-
-  @Override
-  public Query rewrite(IndexReader reader) throws IOException {
-    boolean allReadersRewritable = true;
-    for (LeafReaderContext context : reader.leaves()) {
-      LeafReader leaf = context.reader();
-      Terms terms = leaf.terms(field);
-      PointValues pointValues = leaf.getPointValues(field);
-      if ((terms == null || terms.getDocCount() != leaf.maxDoc())
-          && (pointValues == null || pointValues.getDocCount() != leaf.maxDoc())) {
-        allReadersRewritable = false;
-        break;
-      }
-    }
-    if (allReadersRewritable) {
-      return new MatchAllDocsQuery();
-    }
-    return super.rewrite(reader);
-  }
-
-  @Override
-  public Weight createWeight(IndexSearcher searcher, ScoreMode scoreMode, float boost) {
-    return new ConstantScoreWeight(this, boost) {
-      @Override
-      public Scorer scorer(LeafReaderContext context) throws IOException {
-        DocIdSetIterator iterator = getDocValuesDocIdSetIterator(field, context.reader());
-        if (iterator == null) {
-          return null;
-        }
-        return new ConstantScoreScorer(this, score(), scoreMode, iterator);
-      }
-
-      @Override
-      public int count(LeafReaderContext context) throws IOException {
-        final LeafReader reader = context.reader();
-        final FieldInfo fieldInfo = reader.getFieldInfos().fieldInfo(field);
-        if (fieldInfo == null || fieldInfo.getDocValuesType() == DocValuesType.NONE) {
-          return 0; // the field doesn't index doc values
-        } else if (reader.hasDeletions() == false) {
-          if (fieldInfo.getPointDimensionCount() > 0) {
-            return reader.getPointValues(field).getDocCount();
-          } else if (fieldInfo.getIndexOptions() != IndexOptions.NONE) {
-            return reader.terms(field).getDocCount();
-          }
-        }
-        return super.count(context);
-      }
-
-      @Override
-      public boolean isCacheable(LeafReaderContext ctx) {
-        return DocValues.isCacheable(ctx, field);
-      }
-    };
+    super(field);
   }
 
   /**
