+++ lucene/core/src/java/org/apache/lucene/search/KnnVectorFieldExistsQuery.java
@@ -17,65 +17,17 @@
 
 package org.apache.lucene.search;
 
-import java.io.IOException;
-import java.util.Objects;
-import org.apache.lucene.index.LeafReaderContext;
-
 /**
  * A {@link Query} that matches documents that contain a {@link
  * org.apache.lucene.document.KnnVectorField}.
+ *
+ * @deprecated Use {@link org.apache.lucene.search.FieldExistsQuery} instead.
  */
-public class KnnVectorFieldExistsQuery extends Query {
-
-  private final String field;
+@Deprecated
+public class KnnVectorFieldExistsQuery extends FieldExistsQuery {
 
   /** Create a query that will match documents which have a value for the given {@code field}. */
   public KnnVectorFieldExistsQuery(String field) {
-    this.field = Objects.requireNonNull(field);
-  }
-
-  public String getField() {
-    return field;
-  }
-
-  @Override
-  public boolean equals(Object other) {
-    return sameClassAs(other) && field.equals(((KnnVectorFieldExistsQuery) other).field);
-  }
-
-  @Override
-  public int hashCode() {
-    return 31 * classHash() + field.hashCode();
-  }
-
-  @Override
-  public String toString(String field) {
-    return "KnnVectorFieldExistsQuery [field=" + this.field + "]";
-  }
-
-  @Override
-  public void visit(QueryVisitor visitor) {
-    if (visitor.acceptField(field)) {
-      visitor.visitLeaf(this);
-    }
-  }
-
-  @Override
-  public Weight createWeight(IndexSearcher searcher, ScoreMode scoreMode, float boost) {
-    return new ConstantScoreWeight(this, boost) {
-      @Override
-      public Scorer scorer(LeafReaderContext context) throws IOException {
-        DocIdSetIterator iterator = context.reader().getVectorValues(field);
-        if (iterator == null) {
-          return null;
-        }
-        return new ConstantScoreScorer(this, score(), scoreMode, iterator);
-      }
-
-      @Override
-      public boolean isCacheable(LeafReaderContext ctx) {
-        return true;
-      }
-    };
+    super(field);
   }
 }
